// plc-industrial-ui/src/app/api/model/plc-reading-event.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface PlcReadingEvent { 
    source?: string;
    tagName?: string;
    address?: string;
    dataType?: string;
    valueAsString?: string;
    timestamp?: string;
    quality?: PlcReadingEvent.QualityEnum;
}
export namespace PlcReadingEvent {
    export const QualityEnum = {
        Good: 'GOOD',
        Bad: 'BAD',
        Uncertain: 'UNCERTAIN'
    } as const;
    export type QualityEnum = typeof QualityEnum[keyof typeof QualityEnum];
}